Do not set USER in Docker ENV, use login shell
As reported in #4, setting the `USER` env var in Docker/Singularity
can result in clashes on multiuser systems due to these sharing the
`snemo` value (but not UID).

Remove setting of `USER` in the Docker ENV command. Use login shell
generally and in ENTRYPOINT to ensure this variable is set to the
actual user in the container.

Update README with corrected info for Singularity, enabling easier
use of `brew snemo-shell` in Singularity.

Fixes: #4